@@ -1,47 +1,53 @@
 {
-  "name": "node-js-getting-started",
-  "version": "0.3.0",
-  "description": "A sample Node.js app using Express 4",
+  "name": "danilo-aleixo-portfolio",
+  "version": "1.0.0",
+  "description": "Danilo Aleixo's Personal Portfolio Website",
+  "type": "module",
   "engines": {
-    "node": "8.9.1"
+    "node": ">=18.0.0"
   },
   "main": "index.js",
   "scripts": {
     "start": "node index.js",
+    "dev": "nodemon index.js",
+    "build": "gulp build",
+    "watch": "gulp watch",
     "test": "node test.js"
   },
   "dependencies": {
-    "@sendgrid/mail": "^6.1.4",
-    "body-parser": "^1.18.2",
-    "cool-ascii-faces": "1.3.4",
-    "ejs": "^2.5.6",
-    "express": "^4.15.2"
+    "@sendgrid/mail": "^8.1.1",
+    "body-parser": "^1.20.2",
+    "cool-ascii-faces": "^1.3.4",
+    "ejs": "^3.1.9",
+    "express": "^4.18.2"
   },
   "devDependencies": {
-    "gulp": "^3.9.1",
-    "gulp-autoprefixer": "^3.1.1",
-    "gulp-plumber": "^1.1.0",
-    "gulp-rename": "^1.2.2",
-    "gulp-sass": "^2.3.2",
-    "gulp-uglify": "^2.0.0",
-    "node-sass": "^4.5.3",
-    "node-env-run": "^2.0.1",
-    "request": "^2.81.0",
-    "tape": "^4.7.0"
+    "gulp": "^4.0.2",
+    "gulp-autoprefixer": "^8.0.0",
+    "gulp-plumber": "^1.2.1",
+    "gulp-rename": "^2.0.0",
+    "gulp-sass": "^5.1.0",
+    "gulp-uglify": "^3.0.2",
+    "sass": "^1.69.5",
+    "nodemon": "^3.0.2",
+    "node-fetch": "^3.3.2",
+    "tape": "^5.6.1"
   },
   "repository": {
     "type": "git",
-    "url": "git+https://github.com/heroku/node-js-getting-started.git"
+    "url": "git+https://github.com/daniloaleixo/daniloaleixo.github.io.git"
   },
   "keywords": [
+    "portfolio",
+    "personal",
+    "website",
     "node",
-    "heroku",
     "express"
   ],
   "license": "MIT",
   "bugs": {
-    "url": "https://github.com/heroku/node-js-getting-started/issues"
+    "url": "https://github.com/daniloaleixo/daniloaleixo.github.io/issues"
   },
-  "homepage": "https://github.com/heroku/node-js-getting-started#readme",
-  "author": ""
+  "homepage": "https://github.com/daniloaleixo/daniloaleixo.github.io#readme",
+  "author": "Danilo Aleixo"
 }
