@@ -1,39 +1,50 @@
-var gulp = require('gulp');
-var plumber = require('gulp-plumber');
-var uglify = require('gulp-uglify');
-var sass = require('gulp-sass');
-var rename = require('gulp-rename');
-var autoprefixer = require('gulp-autoprefixer');
+import gulp from 'gulp';
+import plumber from 'gulp-plumber';
+import uglify from 'gulp-uglify';
+import sass from 'gulp-sass';
+import * as dartSass from 'sass';
+import rename from 'gulp-rename';
+import autoprefixer from 'gulp-autoprefixer';
 
-gulp.task('scripts', function() {
+const sassCompiler = sass(dartSass);
+
+// Scripts task
+function scripts() {
     return gulp.src('public/js/scripts.js')
-        .pipe(plumber(plumber({
+        .pipe(plumber({
             errorHandler: function (err) {
                 console.log(err);
                 this.emit('end');
             }
-        })))
-        .pipe(uglify({
-            preserveComments: 'license'
         }))
+        .pipe(uglify())
         .pipe(rename({extname: '.min.js'}))
         .pipe(gulp.dest('public/js'));
-});
+}
 
-gulp.task('styles', function() {
+// Styles task
+function styles() {
     return gulp.src('scss/styles.scss')
-        .pipe(plumber(plumber({
+        .pipe(plumber({
             errorHandler: function (err) {
                 console.log(err);
                 this.emit('end');
             }
-        })))
-        .pipe(sass({outputStyle: 'compressed'}))
+        }))
+        .pipe(sassCompiler({outputStyle: 'compressed'}))
         .pipe(autoprefixer())
         .pipe(gulp.dest('public/css'));
-});
+}
 
-gulp.task('watch', ['scripts', 'styles'], function() {
-    gulp.watch('public/js/*.js', ['scripts']);
-    gulp.watch('scss/*.scss', ['styles']);
-});
+// Watch task
+function watch() {
+    gulp.watch('public/js/*.js', scripts);
+    gulp.watch('scss/*.scss', styles);
+}
+
+// Build task
+const build = gulp.parallel(scripts, styles);
+
+// Export tasks
+export { scripts, styles, watch, build };
+export default build;
